Color modifier now uses byte instead of int
I wanted to originally give negative values to this, but I realized I could override the minus operator instead and subject positive values, rather than adding negative values.

+++ LineInput/Color.cs
@@ -15,7 +15,7 @@ namespace LineInput
             B = b;
         }
 
-        public static Red Red(int value) => new Red(value);
+        public static Red Red(byte value) => new Red(value);
 
         public override bool Equals(object obj)
         {
